+++ core-ui/src/components/App/App.js
@@ -1,9 +1,7 @@
-import React, { useEffect, useMemo, Suspense } from 'react';
+import React, { useEffect, useMemo } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import i18next from 'i18next';
 
-import { Spinner } from 'shared/components/Spinner/Spinner';
 import { MainFrameRedirection } from 'shared/components/MainFrameRedirection/MainFrameRedirection';
 import { useMicrofrontendContext } from 'shared/contexts/MicrofrontendContext';
 import { WithTitle } from 'shared/hooks/useWindowTitle';
@@ -16,6 +14,7 @@ import { useResourceSchemas } from './resourceSchemas/useResourceSchemas';
 import { AppContext } from './AppContext';
 
 import { resourceRoutes } from 'resources';
+import { createExtensibilityRoutes } from './ExtensibilityRoutes';
 import otherRoutes from 'resources/other';
 
 export default function App() {
@@ -60,74 +59,9 @@ export default function App() {
           }
         />
 
-        {customResources?.map(cr => {
-          const List = React.lazy(() =>
-            import('../Extensibility/ExtensibilityList'),
-          );
-          const Details = React.lazy(() =>
-            import('../Extensibility/ExtensibilityDetails'),
-          );
-
-          const translationBundle = cr?.resource?.path || 'extensibility';
-          i18next.addResourceBundle(
-            language,
-            translationBundle,
-            cr?.translations?.[language] || {},
-          );
-
-          if (cr.resource?.scope === 'namespace') {
-            return (
-              <React.Fragment key={`namespace-${cr.resource?.path}`}>
-                <Route
-                  path={`/namespaces/:namespaceId/${cr.resource.path}`}
-                  exact
-                  element={
-                    <Suspense fallback={<Spinner />}>
-                      <List />
-                    </Suspense>
-                  }
-                />
-                {cr.details && (
-                  <Route
-                    path={`/namespaces/:namespaceId/${cr.resource.path}/:resourceName`}
-                    element={
-                      <Suspense fallback={<Spinner />}>
-                        <Details />
-                      </Suspense>
-                    }
-                  />
-                )}
-              </React.Fragment>
-            );
-          } else {
-            return (
-              <React.Fragment key={`cluster-${cr.resource?.path}`}>
-                <Route
-                  path={`/${cr.resource.path}`}
-                  exact
-                  element={
-                    <Suspense fallback={<Spinner />}>
-                      <List />
-                    </Suspense>
-                  }
-                />
-                {cr.details && (
-                  <Route
-                    path={`/${cr.resource.path}/:resourceName`}
-                    element={
-                      <Suspense fallback={<Spinner />}>
-                        <Details />
-                      </Suspense>
-                    }
-                  />
-                )}
-              </React.Fragment>
-            );
-          }
-        })}
-
         {resourceRoutes}
         {otherRoutes}
+        {customResources?.map(cr => createExtensibilityRoutes(cr, language))}
         <Route path="" element={<MainFrameRedirection />} />
       </Routes>
     </AppContext.Provider>
